Change short to int
I am questioning my life. Why does an int have better performance that a short?

+++ ConsoleApplication3/Constants.h
@@ -3,9 +3,9 @@ class Constants
 {
 public:
 	struct RGBcolor {
-		short red = 0;
-		short green = 0;
-		short blue = 0;
+		int red = 0;
+		int green = 0;
+		int blue = 0;
 	};
 
 	RGBcolor NONE;
